Manually tell the kubelet its IP
For some reason each kubelet sets its IP to the same value. The cluster
appears to work but then kubectl cannot access logs or run command on
deployed containers. Hardcoding the IP for the kubelet like this
appears to fix the problem without having the internal knowledge of
what the kubelet is trying to do in the first place.